animatie afgerond, overgangen lopen soepel. To do: alert on refresh. Convert to .pdf?

+++ src/pages/login/Login.js
@@ -3,42 +3,48 @@ import './Login.scss'
 import Background from "../../components/Background/Background";
 import Input from "../../components/Input/Input"
 import {Button} from "../../components/Button/Button";
-import {useNavigate, Navigate} from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import axios from 'axios'
-import {AuthContext} from "../../context/AuthContext";
+import {AuthContext} from "../../context/AuthContext/AuthContext";
+import {UserContext} from "../../context/UserContext/UserContext";
+import ErrorMessage from "../../components/ErrorMessage/ErrorMessage";
 
 function Login(){
     const {setAuthState, authState} = useContext(AuthContext)
-    const noviBackend = 'https://frontend-educational-backend.herokuapp.com/'
+    const {error, toggleError} = useContext(UserContext)
     const navigate = useNavigate();
-    const [error, toggleError] = useState(false)
+    const [startup, toggleStartup] = useState(true)
     const [userLogin, setUserLogin] = useState({
         username: "",
         password: ""
     })
 
-    useEffect(() => {
+useEffect(() => {
+        if(startup) {
+            setTimeout(() => {
+            toggleStartup(false)
+        }, 1900)
+        }
     async function contactBackend(){
         try{
             const data = await axios.get( 'https://frontend-educational-backend.herokuapp.com/api/test/all')
-            console.log(data)
         }
         catch(e){
             console.error(e)
         }
     }
         contactBackend()
-        localStorage.clear()
-        setAuthState({
-            ...authState,
-            user: null,
-            authStatus: 'done',
-            isAuth: false
-        })
+        if(authState.isAuth === true) {
+            localStorage.clear()
+            setAuthState({
+                ...authState,
+                user: null,
+                authStatus: 'done',
+                isAuth: false
+            })
+        }
         sessionStorage.clear()
     }, [])
-
-
 
     async function handleLogin(){
         try{
@@ -46,18 +52,22 @@ function Login(){
                 "username": userLogin.username,
                 "password": userLogin.password
             })
-            toggleError(false)
-            console.log(data)
+            toggleError({
+                ...error,
+                loginError: false
+            })
             localStorage.setItem('token', data.accessToken)
             setTimeout(() => {
             navigate('/home')}, 500)
         }
         catch(e){
             console.error(e)
-            toggleError(true)
+            toggleError({
+                ...error,
+                loginError: true
+            })
         }
     }
-
 
     function handleChange(e){
         setUserLogin({
@@ -72,71 +82,84 @@ function Login(){
     }
 
     return (
-        <Background
-            classNameTop="background-top"
-            classNameCenter="background-center"
-            classNameBottom="background-bottom"
-            topContent={
+        <>
+        { startup ?
+                <Background
+                    classNameTop="background-top"
+                    classNameCenter="background-center"
+                    classNameBottom="background-bottom"
+                    centerContent={
+                        <div className="startup-logo--h1">
+                            <h1>Setlist</h1>
+                            <h1>Manager</h1>
+                        </div>
+                    }
+                />
+            :
                 <>
-                    <header className="login--header">
-                        <h1 className="setlist">Setlist</h1>
-                        <h1 className="manager">Manager</h1>
-                    </header>
-                </>
-            }
-            centerContent={
-                <>
-                    <div className="loginContainer">
-                        <span className="login">
-                            <h2 className="login__h2">
-                                Login
-                            </h2>
-                        </span>
-                        <form
-                            className="loginForm"
-                            onSubmit={handleSubmit}
-                        >
-                            <Input
-                                label="Username"
-                                className="textInput"
-                                inputType="text"
-                                inputName="username"
-                                onChange={handleChange}
-                            />
-                            <Input
-                                label="Password"
-                                className="textInput"
-                                inputType="password"
-                                inputName="password"
-                                onChange={handleChange}
-                            />
-                            {error &&
-                            <p className="login-error-message">
-                                *Username & password combination is incorrect, please try again or register first.
-                            </p>}
+                <Background
+                    classNameTop="background-top"
+                    classNameCenter="background-center"
+                    classNameBottom="background-bottom"
+                    topContent={
+                            <header className="login--header">
+                                <h1 className="setlist">Setlist</h1>
+                                <h1 className="manager">Manager</h1>
+                            </header>
+                    }
+                    centerContent={
+                        <div className="loginContainer">
+                            <span className="login">
+                                <h2 className="login__h2">
+                                    Login
+                                </h2>
+                            </span>
+                            <form
+                                className="loginForm"
+                                onSubmit={handleSubmit}
+                            >
+                                <Input
+                                    label="Username"
+                                    className="textInput"
+                                    inputType="text"
+                                    inputName="username"
+                                    onChange={handleChange}
+                                />
+                                <Input
+                                    label="Password"
+                                    className="textInput"
+                                    inputType="password"
+                                    inputName="password"
+                                    onChange={handleChange}
+                                />
+                                <ErrorMessage
+                                    loginError={error.loginError}
+                                />
                                 <Button
-                                type="submit"
-                                className="loginForm__button"
-                                buttonText="Login"
+                                    type="submit"
+                                    className="loginForm__button"
+                                    buttonText="Login"
                                 />
                             </form>
                         </div>
-                    </>
                     }
                     bottomContent={
-                    <>
-                        <Button
-                            type="button"
-                            className="registerButton--start-register"
-                            onClick={() => {
-                                navigate("/register")
-                            }}
-                            buttonText="Or Register"
-                        />
-                    </>
-                }
-                    />
-            )
+                        <footer className="login--footer">
+                            <Button
+                                type="button"
+                                className="registerButton--start-register"
+                                onClick={() => {
+                                    navigate("/register")
+                                }}
+                                buttonText="Or Register"
+                            />
+                        </footer>
+                    }
+                />
+            </>
+        }
+    </>
+    )
 }
 
 export default Login
